@@ -1,8 +1,8 @@
 [project]
 name = "configur8"
 description = "Type-safe configuration and validation library"
-authors = [{ name = "Nick Joyce", email = "nick@stratuscode.com" }]
-requires-python = ">=3.7,<4.0"
+authors = [{ name = "Nick Joyce", email = "nick@boxdesign.co.uk" }]
+requires-python = ">=3.8"
 readme = "README.md"
 keywords = [
   "config",
@@ -16,42 +16,17 @@ keywords = [
   "python-config",
 ]
 classifiers = [
-    "Topic :: Software Development",
-    "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.7",
-    "Programming Language :: Python :: 3.8",
-    "Programming Language :: Python :: 3.9",
-    "Programming Language :: Python :: 3.10",
-    "Programming Language :: Python :: 3.11",
+  "Topic :: Software Development",
+  "Programming Language :: Python :: 3.12",
 ]
 license = { text = "MIT" }
 dynamic = ["version"]
 
-dependencies = [
-    "email-validator>=1.3.1",
-    "typing-extensions>=4.0.0",
-]
-
-[tool.pdm.dev-dependencies]
-dev = [
-    "codecov",
-    "flake8",
-    "mypy",
-    "pytest-cov",
-    "pytest-mock",
-    "pytest-mypy-plugins",
-    "pytest",
-    "types-pyyaml",
-    "setuptools>=67.2.0",
-]
-
-[tool.pdm.version]
-source = "file"
-path = "src/configur8/__about__.py"
+dependencies = ["email-validator"]
 
 [build-system]
-requires = ["pdm-pep517>=1.0.0"]
-build-backend = "pdm.pep517.api"
+requires = ["pdm-backend"]
+build-backend = "pdm.backend"
 
 [tool.pytest.ini_options]
 addopts = """
@@ -62,3 +37,26 @@ addopts = """
   --mypy-only-local-stub
   --ignore=__pypackages__
 """
+
+[tool.pdm.version]
+source = "file"
+path = "src/configur8/__about__.py"
+
+
+[dependency-groups]
+dev = [
+  "codecov",
+  "mypy",
+  "pytest-cov",
+  "pytest-mock",
+  "pytest-mypy-plugins",
+  "pytest",
+  "ruff",
+  "types-pyyaml",
+]
+
+[tool.ruff]
+line-length = 80
+
+[tool.ruff.lint]
+select = ["E", "F", "W", "C90"]
