@@ -3,5 +3,5 @@ __all__ = (
     "__version_info__",
 )
 
-__version__ = "2.0.1"
+__version__ = "2.1.0"
 __version_info__ = [int(x) for x in __version__.split(".")]
