@@ -11,9 +11,14 @@ def parse(email: str) -> str:
     except EmailNotValidError as exc:
         raise InvalidConfig(str(exc)) from exc
 
-    if not result.email:
+    if hasattr(result, "normalized"):
+        res = result.normalized
+    else:
+        res = result.email
+
+    if not res:
         raise InvalidConfig(f"Email is not valid?? ({email!r})")
 
-    assert isinstance(result.email, str)
+    assert isinstance(res, str)
 
-    return result.email
+    return res
