@@ -1,14 +1,14 @@
 import enum
 
 __all__ = (
-    "Empty",
+    "Missing",
     "MISSING",
 )
 
 
 # some stupid nonsense to get mypy to work with sentinels
-class Empty(enum.Enum):
+class Missing(enum.Enum):
     token = 0
 
 
-MISSING = Empty.token
+MISSING = Missing.token
